+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Tests for "Unary".
+- Tests for "Temperatures".
 
 ## [1.0.0] - 2022-02-07
 ### Added